refactor(abw): separate composite image workflows
